Refactor AbstractInsn class
