--- compiler/llvmAbstractInsnGen.cpp
@@ -1,6 +1,0 @@
-#include "BIR.h"
-
-AbstractInsn::AbstractInsn(Location *pos, InstructionKind kindP, Operand *lOp)
-    : kind(kindP), lhsOp(lOp) {
-  setLocation(pos);
-}
